Fix possible null dereferences in ScimUserBootstrap
AuthEvent handling might result in null dereference if fetch of SCIM
user fails when trying to update SCIM user. If SCIM user fetch fails,
and the user is null, throw an explicit RuntimeException as is done in
another handler, rather than have a NullPointerException get thrown when
the null user is dereferenced.