Fix possible null dereferences in ScimUserBootstrap
AuthEvent handling might result in null dereference if fetch of SCIM
user fails when trying to update SCIM user. Check for user retrieved not
being null before performing update. Previously if the retrieval failed
and the user was null, a null pointer exception would have been thrown.
With this change, the update will fail silently, but the failed
retrieval will be logged as before.