Use transient artifactory location for non-release uploads
If the parameter to upload artifacts to artifactory is checked when
building, the upload happens currently to MAAXA folder that retains
artifacts forever. Instead, for builds from branches other than release
branches, push to transient MAAXA-SNAPSHOT folder as is done by other
projects. This also prevents somebody accidentally uploading artifacts
from a PR build and overwriting a release build's artifacts, as the
location is based on the version set in gradle.properties.

- Create a new common file shared between build and deploy Jenkinsfiles
  that has functions to get application version, and to get upload path
  in artifactory, based on the branch name.
- Use `MAAXA/builds/uaa/<version>/` folder as before for build/deploy of
  release_xxxx branches
- Use `MAAXA-SNAPSHOT/builds/uaa/<branch-name>` for build/deploy of
  non-release branches

With this change, it is possible to trigger a branch build job (branch
or PR) with the parameter to upload artifacts enabled, and then trigger
a branch deploy job which will automatically pick the artifact uploaded
by the corresponding branch build job (based on branch/PR name).