Update all Jenkins pipelines to keep 30 builds
Increase number of builds and artifacts (test reports) to keep from 15
to 30, and remove time based removal of builds. Some release branches
will not get built for months, but it is still useful to have build
history for them. With a time limit, build history for such branches is
lost.