Disable SAML login acceptance test for locally run UAA
The SAML test requires UAA to be registered with GE SSO as a SAML app
[1]. This has been done for CF environments, but not for locally run
UAA, so the test will fail for local UAA, whether that be a native Java
run or a containerized run. Currently the acceptance tests are being run
for dockerized UAA in the DockerizeJenkins pipeline [2], which is
failing as it is also a 'local'container run on Jenkins, without any
deployment to Kubernetes. Fix the failure by disabling the test for
local UAA runs.

- Update acceptance test task in build.gradle to exclude SAML Login test
  if the `DEPLOYMENT_TYPE` is `local`. This is the case for the
  dockerized UAA Jenkins pipeline [3].

[1] https://devcloud.swcoe.ge.com/devspace/display/RMOCM/SamlLoginAT+-+How+to+update+metadata+for+GESSO
[2] https://github.com/GESoftware-CF/uaa/blob/release_75.6.0/DockerizeJenkinsfile#L75
[3] https://github.com/GESoftware-CF/uaa/blob/release_75.6.0/DockerizeJenkinsfile#L62