Update method to specify UAA URL for JWT Bearer grant tests
Instead of specifying UAA URL as a combination of UAA_HOST:UAA_PORT, use
UAA_ROUTE, which is also exported by the test running script. This
simplifies things, and lets the test script pick the route if specified
in the environment. This is useful for local minikube runs, which like
local runs is tested over HTTP, but where the UAA URL does not have a
port specified, and no CF steps are needed to be executed (see
uaa-cf-release PR [1])

[1] https://github.build.ge.com/predix/uaa-cf-release/pull/58