[jwt-grant] Update method to specify UAA URL for JWT Bearer grant tests
Instead of specifying UAA URL as a combination of UAA_HOST:UAA_PORT, use
UAA_ROUTE, which is also exported by the test running script. This
simplifies things, but lets the test script pick the route if specified
in the environment, for local runs. This is useful for local minikube
runs where the test is over HTTP, and no CF steps are needed to be
executed (see uaa-acceptance-test.sh [1])

[1] https://github.build.ge.com/predix/uaa-cf-release/blob/master/uaa-acceptance-tests.sh