[predix-extensions] k8s: Set SECRETS_DIR as /uaa/secrets for docker images
Update Dockerfile to create /uaa/secrets and set SECRETS_DIR environment
variable to point to it. Any yml files in the directory will get parsed
by UAA on startup [1]. This makes it possible to have 'secret'
configuration be split across yml files and mounted into the container
as Secrets as is done in k8s/templates/secrets.

[1] https://github.com/GESoftware-CF/uaa/commit/82a1072bf6c092e4d04ce20a5f80068cf009f969