Fix taking browser screenshot on UI test failure
In v1.4.1.RELEASE, Spring Boot fixed the issue with selenium WebDriver
lifecycle [1]. According to docs [2], by default, Spring Boot puts
WebDriver beans in a special scope to ensure that the driver exits
after each test and that a new instance is injected. This fix broke the
functionality of taking a screenshot on failure of selenium UI tests -
Spring Boot implements WebDriverTestExecutionListener [3] and calls
WebDriver quit() after each test before ScreenshotOnFail rule is
executed.

Fix screenshot capture by implementing custom TestExecutionListener and
invoking it before control is passed to WebDriverTestExecutionListener
and WebDriver is closed.

- Change implementation of ScreenshotOnFail from TestWatcher to
TestExecutionListener.
- Position ScreenshotOnFail listener so that it‘s invoked before
control is passed to WebDriverTestExecutionListener and WebDriver is
closed. Listeners execution is driven from lowest order to highest for
the beforeTestMethod callback and in the reverse order for the
afterTestMe [4]
- Register ScreenshotOnFail listener using @TestExecutionListeners
annotation in each selenium test class. MERGE_WITH_DEFAULTS indicates
that the locally declared listeners should be merged with the default
Spring listeners.
- Intentionally didn’t register ScreenshotOnFail listener in
`META-INF/spring.factories` to avoid automatic discovery by the
SpringFactoriesLoader mechanism and to prevent invocation of
ScreenshotOnFail for non-selenium tests.

[1] https://github.com/spring-projects/spring-boot/commit/ac2609b5857727724709e7da793d82264bd6726d
[2] https://docs.spring.io/spring-boot/docs/2.6.6/reference/html/features.html#features.testing.spring-boot-applications.spring-mvc-tests
[3] https://github.com/spring-projects/spring-boot/blob/v2.6.6/spring-boot-project/spring-boot-test-autoconfigure/src/main/java/org/springframework/boot/test/autoconfigure/web/servlet/WebDriverTestExecutionListener.java
[4] https://www.baeldung.com/spring-testexecutionlistener