Add mandatory parameters field to orchestrator zone response
Per orchestrator documentation [1] `parameters` field is required in
service response JSON only if service supports update. However, current
orchestrator implementation depends on parameters to be present and set
to NULL for services that do not support update.

- Add `parameters` field to OrchestratorZoneResponse model. Removing
`Include.NON_NULL` annotation [2] allows it to be included in the
response with null value

[1] https://devcloud.swcoe.ge.com/devspace/pages/viewpage.action?spaceKey=PCE&title=GED+Services+-+Orchestrator+Integration
[2] https://github.com/GESoftware-CF/uaa/pull/558/commits/2a3ff845c2882f17442ddab4656505b9991cb8a4