Add Spring AWS Secret Manager dependencies
Add spring aws cloud secret manager dependency to build.gradle, also add
aws sdk sts. Since this is not a proper spring boot application, load
the autoconfiguration in the servlet xml.

All secrets are stored in yaml format in aws secret manager using `Plaintext`
view[1] and uaa is updated to create in-memory ByteArrayResource from the same
in yaml format and added to the existing resources list in `YamlServletProfileInitializer.java`
for the parsing by existing YamlProcessor.

[1] https://github.devtools.predix.io/predix-security/uaa-k8s-deploy/pull/57