Create client with jwt bearer with empty secret (#3301)
* Refactoring to fix issue #3232

Move the check for existing secrets to the
grant flow (runtime) and remove it from REST (configuration).

We support now private_key_jwt next to secret, so check all parts in REST is too less and will get complicated.
Fix a potential security issue because we must not create a token based on no authentication

* Tests to allow client creation without secret checks

* comment

* rename test

* revert