feature test: JWT bearer across zones (#3348)
* fix: Jwt Bearer should be able to return id_token

I think it was simply forgotten to allow it, but
in documentation there is response_type with "token id_token", but the flow is not allowed to return it.
Fix this

* feature test: Test JWT bearer across zones

The test uses 3 zones in UAA to represent
3 identity zones.

This PR needs https://github.com/cloudfoundry/uaa/pull/3344
because jwt bearer should / must return an id_token if the client is configured with openid, similar to password grant.

* setup proxy test